requirements.txt
streamlit
pandas
networkx
pillow
openai
scipy
jsonschema
